requirements.txt
accelerate==1.13.0
annotated-doc==0.0.4
anyio==4.13.0
certifi==2026.2.25
click==8.3.2
colorama==0.4.6
filelock==3.29.0
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.12
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
packaging==26.1
pandas==3.0.2
psutil==7.2.2
Pygments==2.20.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2026.4.4
rich==15.0.0
safetensors==0.7.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.5.4
typer==0.24.1
typing_extensions==4.15.0
tzdata==2026.1
